[sailfish-office] Improve responsiveness for non-trivial PDFs.
Though the PDFRenderThread used a QQueue, it was implemented
to block the second page request until the first one was
completed, so it could at most render one page at a time before
being blocked. Changed the implementation to be a fully
functioning queue of rendering requests.

The page render thread was ticked by a timer which was started, but
never stopped, so this would tick every 50 milliseconds, and never
stop, regardless of if the application needed to render or
not. Changed this to post an event to the page render thread after
adding a request so that we go dormant when nothing is pending.

The logic was set up to make requests in a fifo manner, so when
we scrolled far or zoomed in, a lot of outdated pages would be
rendered before we got to our visible page. Worst case, I had 255
page requests each taking 4 seconds to render. Change this to
remove all pending requests on layout changes or when scrolling
beyond the loaded area.

When a page is requested and should be shown on screen, we try
to prioritize it by moving it to the front of the job queue. This
reduces the amount of time spent waiting with a blank screen
when scrolling beyond the loaded area when pages are slow to
render.

The textures are created on the page rendering thread which makes it
possible for the scene graph to use fully asynchronous textures
(requires: https://codereview.qt-project.org/#change,86146 or similar)
which means that we will not skip frames while panning up and down,
also when new zoomed-in pages are uploaded.

We could still improve this quite a bit by splitting the page into
tiles but this change should go a long way towards making the
application usable.