[office] Add translation support for some Calligra strings. Fixes JB#45824
Would be nice if Calligra had its translations available, but they
are pain to work with, being located in huge kde subversion repository,
away from the main code. Also everything included would mean
tens of megabytes of data when only a few string are ever shown
to the user on sailfish-office. And even then we'd be forced to show
what Calligra decided as format and available translations.

So let's just hack the few known strings separately for now.