[sailfish-office] Reduce the frequency of PDF tile updates. Fixes JB#49269
Don't cancel all requests on a resize as many events come through at
once and it defeats delay from the resize timer. And move the request
scheduling from updatePaintNode() to updatePolish() which can be called
less frequently and signals can safely emitted from as its called from
the UI thread and not the render thread.