[sailfish-office] Adapt sparql to Tracker3. Contributes to JB#53968 OMP#JOLLA-74
Plain tracker driver was implemented directly on top of d-bus. Better
use the "direct" version which, somewhat confusingly based on name,
is implemented with tracker-sparql library.

On the query there are now multiple graphs and we are mostly interested
on tracker:Documents.

The file types of interest are all having PaginatedTextDocument so
skipping the separated list. Assuming tracker-miner rules adjusted
so that rtf and csv are being inserted into the Documents graph.

tracker:available is no longer available for items directly,
instead need to check that via its dataSource.

nie:isStoredAs is marked deprecated in Tracker, but yet
it's used even on new example code and Tracker3 ports of
other projects.

FileSize now fetched from the Documents graph, but the PR I have
upstream for adding the info there is not yet merged.

LastAccessed changed to lastModified. Accessed is not available at the
moment on the Documents graph and in general the atime is not commonly
getting updated either on filesystem nor on tracker database.